improve workflow symbol docs
