+++ .github/workflows/shared.tsx
@@ -1,25 +1,17 @@
+/**
+ * CheckoutStep is a step that checks out the repository.
+ */
 export function CheckoutStep() {
   return {
     uses: "actions/checkout@v4",
   };
 }
 
+/**
+ * SetupDenoStep is a step that sets up the Deno runtime.
+ */
 export function SetupDenoStep() {
   return {
     uses: "denoland/setup-deno@v1",
-  };
-}
-
-export function DenoFormatStep() {
-  return {
-    name: "Format",
-    run: "deno fmt && git diff-index --quiet HEAD",
-  };
-}
-
-export function DenoLintStep() {
-  return {
-    name: "Lint",
-    run: "deno lint && git diff-index --quiet HEAD",
   };
 }
